Rename package directory to faster_llm
