+++ .travis.yml
@@ -8,7 +8,7 @@ install:
   - pip install codecov # Install coveralls
 
 script:
-  - pytest tests --cov=faster_ds # Run unittest
+  - pytest tests --cov=faster_llm # Run unittest
 
 notifications:
   email: false # Mute email notification
