+++ LICENSE.txt
@@ -1,6 +1,6 @@
 The MIT License (MIT)
 
-Copyright 2019 fuwiak <https://github.com/fuwiak/faster_ds>
+Copyright 2019 fuwiak <https://github.com/fuwiak/faster_llm>
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
