+++ docs/source/index.rst
@@ -1,9 +1,9 @@
-.. faster_ds documentation master file, created by
+.. faster_llm documentation master file, created by
    sphinx-quickstart on Thu Apr  9 18:12:15 2020.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to faster_ds's documentation!
+Welcome to faster_llm's documentation!
 =====================================
 
 .. toctree::
