+++ faster_llm/doc.py
