+++ tox.ini
@@ -14,4 +14,4 @@ deps =
 
 commands =
     pip install --editable .
-    pytest tests --cov=faster_ds
+    pytest tests --cov=faster_llm
